deploy args / mysql  enhancement

+++ deploy/deploy_fedora.sh
@@ -2,17 +2,53 @@
 #deploy OpenAttestation 2.x
 
 ### configuration ###
-oat_server_ip="___ IP ___"
-mysql_password="______ MySQL PASSWD ______"
-saml_password="samlpasswd2"
-portal_password="portalpasswd2"
-p12_password="p2"
-java_home="/usr/lib/jvm/java-7-openjdk-amd64/"
-log_dir="/var/log/tomcat"
+mysql_password=${mysql_password:-""}
+oat_server=${oat_server:-""}
+saml_password=${saml_password:-samlpasswd2}
+portal_password=${portal_password:-portalpasswd2}
+p12_password=${p12_password:-p2}
+log_dir=${log_dir:-/var/log/tomcat}
+
+example()
+{
+    echo $"$0: Usage: $0 --mysql <MySQL passwd of oat server> --ip <IP of oatserver>}"
+}
+
+if [ $# -lt 3 ]; then
+    example
+    exit 1
+fi
+
+while [ $# -gt 1 ]; do
+  case $1 in
+    --mysql)
+        if [ "$2" != "--ip" ];then
+            mysql_password=$2
+            shift 2
+        else
+            shift 1
+        fi
+        ;;
+    --ip)
+        oat_server=$2
+        shift 2
+        ;;
+    *)
+        example
+        exit 1;;
+  esac
+done
+
+echo "oat_server = $oat_server"
+echo "mysql_password = $mysql_password"
+
+if [ "$oat_server" == "" ]; then
+    example
+    exit 1
+fi
 
 ###########
 TOP_DIR=$(cd .. && pwd)
-export JAVA_HOME=$java_home
 conf_dir=${conf_dir:-/etc/intel/cloudsecurity}
 [ -d $conf_dir ] && rm -rf $csonf_dir
 install -d  $conf_dir
@@ -24,10 +60,23 @@ oat_home_dir=${oat_home_dir:-$HOME/.oat}
 install -d  $oat_home_dir
 chown -R tomcat:tomcat $oat_home_dir
 
-###MySQL ### 
-mysql -uroot -p$mysql_password -e 'create database mw_as';
-mysql -uroot -p$mysql_password mw_as < \
-$TOP_DIR/database/mysql/src/main/resources/com/intel/mtwilson/database/mysql/mtwilson.sql
+###MySQL ###
+mysql_status="`netstat -vulntp |grep -i mysqld`"
+if [ -n "$mysql_status" ];then
+    echo "mysqld service is running .."
+else
+    echo "starting mysqld service .."
+    service mysqld start
+fi
+if [ "$mysql_password" == "" ];then
+    mysql -uroot -e 'create database mw_as';
+    mysql -uroot mw_as < \
+    $TOP_DIR/database/mysql/src/main/resources/com/intel/mtwilson/database/mysql/mtwilson.sql
+else
+    mysql -uroot -p$mysql_password -e 'create database mw_as';
+    mysql -uroot -p$mysql_password mw_as < \
+    $TOP_DIR/database/mysql/src/main/resources/com/intel/mtwilson/database/mysql/mtwilson.sql
+fi
 
 cd $conf_dir
 mtwilson_properties()
@@ -126,9 +175,9 @@ attestation_properties
 privacyca_properties
 privacyca_client_properties
 
-sed -i "s/OATAPPR_IP/$oat_server_ip/g" $conf_dir/mtwilson.properties
-sed -i "s/OATAPPR_IP/$oat_server_ip/g" $conf_dir/attestation-service.properties
-sed -i "s/OATAPPR_IP/$oat_server_ip/g" $conf_dir/privacyca-client.properties
+sed -i "s/OATAPPR_IP/$oat_server/g" $conf_dir/mtwilson.properties
+sed -i "s/OATAPPR_IP/$oat_server/g" $conf_dir/attestation-service.properties
+sed -i "s/OATAPPR_IP/$oat_server/g" $conf_dir/privacyca-client.properties
 sed -i "s/MYSQL_PASSWD/$mysql_password/g" $conf_dir/mtwilson.properties
 sed -i "s/PORTAL_PASSWD/$portal_password/g" $conf_dir/trust-dashboard.properties
 sed -i "s/PORTAL_PASSWD/$portal_password/g" $conf_dir/whitelist-portal.properties
